Add base logic for the three main commands;
This patch adds the three basic commands of

- `start`
- `stop`
- `data`

These subcommands will be used to communicate with the SQLite DB while
also passing in references. The flags and options have been tested
manually and `clap-rs` and `rust` make working with this so much easier.

```shell
tick 0.2.0
A time-tracking cli tool that generates useful reports for the last eight days.

USAGE:
    tick [FLAGS] [OPTIONS] [SUBCOMMAND]

FLAGS:
    -h, --help       Prints help information
    -V, --version    Prints version information
    -v, --verbose    Sets the level of verbosity.

OPTIONS:
    -e, --entry <entry>    Set a message for this timer.
    -n, --name <name>      Set a name for this timer.

SUBCOMMANDS:
    data     Export data from timers.
    help     Prints this message or the help of the given subcommand(s)
    start    Start a ticking timer.
    stop     Stop a ticking timer.
```