Move modules and fix borked build;
This patch moves the database module to the relay module as I am
essentially building a very basic ORM with hard-coded queries.
Therefore, it's a bit redundant to call it `database` so I decided to
call it `relay`. This patch also fixes the broken build by recreating
the program using Structs, Impl, and Lifetimes. Learning a lot about
systems programming!