Add save functionality; :clap:
This helps put a major dent in diesel-rs/diesel#376. I was getting stuck
here most likely because of what's remaining on diesel-rs/diesel#130.
I'm not sure what differences SQLite has to Postgresql in terms of not
being about to use `get_result`, but the guide for diesel-rs/diesel#376
should either address the differences by suggesting you use `execute`
instead. Although, I'm not sure how to reproduce the functionality for
`get_result` using SQLite? Maybe a quick call to the database to
retrieve the latest one? We wouldn't know the id, so finding items in
the database would be almost impossible or at best super hacky.