Update Tick to leverage a config.yaml;
This patch adds support for a config.yaml to live at ~/.config/tick/ on
the user's system. This is also where the database now lives. Migrating
a database to this new file location is as easy as moving or copying the
file over from where it was configured previously with
TICK_DATABASE_FILE or DATABASE_URL in the previous releases. This way no
data is lost between versions. The app will crash if there isn't a
~/.config/tick/config.yaml file and if it isn't configured with a
`database_path` key for where the database should live.